+++ src/views/worktable/search-idle.vue
@@ -14,9 +14,10 @@
       border
       fit
       highlight-current-row
+      :default-sort="{prop: 'businessId', order: 'descending'}"
       style="width: 100%;"
     >
-      <el-table-column label="闲置id" prop="id" sortable="custom" align="center" width="120">
+      <el-table-column label="闲置id" prop="businessId" sortable align="center" width="120">
         <template slot-scope="{row}">
           <span>{{ row.businessId }}</span>
         </template>
@@ -96,10 +97,7 @@ export default {
       listQuery: {
         page: 1,
         limit: 20,
-        importance: undefined,
-        title: undefined,
-        type: undefined,
-        sort: '+id'
+        title: '',
       },
       downloadLoading: false,
       picList:[]
@@ -119,30 +117,31 @@ export default {
         }, 1.5 * 1000)
       })
     },
-    handleModifyStatus(row, status) {
-      this.$message({
-        message: '操作Success',
-        type: 'success'
-      })
-      row.status = status
-    },
-
     handleDelete(row, index) {
-      console.log(row.suitId);
-      deleteByBookSuitId(row.businessId).then(response => {
-        console.log(response)
-        this.$notify({
-        title: '成功',
-        message: response.msg,
-        type: 'success',
-        duration: 2000
-      })
-      // 确定OK再删除
-      this.list.splice(index, 1)
-      }).catch(err => {
-        console.log(err);
-      })
+      // 二次确认
+      this.$confirm('此操作将永久删除该闲置, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteByBookSuitId(row.businessId).then(response => {
+          console.log(response)
+          // 确定OK再删除
+          this.$message({
+          type: 'success',
+          message: response.msg
+        });
+          this.list.splice(index, 1)
+        }).catch(err => {
+          console.log(err);
+        })
 
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        });
+      });
 
     },
     searchBlank() {
